feat: Optionally set concurrency limit
